create a base cae
